Shared WASM wrapper used by both Node and browser
